+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Added option "Position of minimap buttons".
+* Added option "Expand to".